Add files via upload
The objective of this program is to encipher and decipher a .txt file specified by the user. 

It is necessary that the specified file be located in the same directory as Cyphertxt.py.
This program warns the user if the specified file did not exist. This is typically triggered due to a missing file extension.
The user is then prompted to specify the destination filename.
When this is entered, a new .txt file is generated in the directory and written to using the python regular expression "w+". 

The rules to this cipher are simple:
*Obtain the first and second half of the string.
*Through iteration, if the index of the character in the string is even, ignore.
* Otherwise, the character at the index is to be moved to the end of the string.